% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_split.R
\name{create_data_split}
\alias{create_data_split}
\title{Create a calibration (training) / validation data split for PLSR model fitting and testing}
\usage{
create_data_split(
  dataset = NULL,
  approach = NULL,
  split_seed = 123456789,
  prop = 0.8,
  group_variables = NULL
)
}
\arguments{
\item{dataset}{input full PLSR dataset to split into cal/val datasets}

\item{approach}{approach to splitting the dataset. Options: base or dplyr}

\item{split_seed}{random seed to use for splitting data}

\item{prop}{the proportion of data to preserve for calibration (e.g. 0.8) and validation (0.2).
This sets the calibration proportion}

\item{group_variables}{Use factor variables to conduct a stratfied sampling for cal/val}
}
\value{
output_list A list containing the calibration dataset (cal_data)
and validation dataset (val_data)
}
\description{
Create a calibration (training) / validation data split for PLSR model fitting and testing
}
\author{
Julien Lamour, Jeremiah Anderson, Shawn P. Serbin
}
