% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsr_data.R
\docType{data}
\name{ely_plsr_data}
\alias{ely_plsr_data}
\title{Ely et al (2019) example leaf-level PLSR dataset. DOI: https://doi.org/10.1093/jxb/erz061}
\format{
An object of class \code{data.frame} with 178 rows and 1908 columns.
}
\usage{
ely_plsr_data
}
\description{
Ely et al (2019) example leaf-level PLSR dataset. DOI: https://doi.org/10.1093/jxb/erz061
}
\keyword{datasets}
