% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.coef.valid.R
\name{f.coef.valid}
\alias{f.coef.valid}
\title{f.coef.valid}
\usage{
f.coef.valid(plsr.out, data_plsr, ncomp, inVar)
}
\arguments{
\item{plsr.out}{plsr model obtained with jaccknife = TRUE}

\item{data_plsr}{data used for the plsr model with Spectra the matrix of spectra}

\item{ncomp}{number of selection components}

\item{inVar}{Name of the PLSR model response variable}
}
\value{
B returns the intercept and the coefficients of the jackknife or bootstrap validation
}
\description{
f.coef.valid
}
\author{
Julien Lamour
}
