% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.plot.coef.R
\name{f.plot.coef}
\alias{f.plot.coef}
\title{f.plot.coef}
\usage{
f.plot.coef(
  Z,
  wv,
  xlim = NULL,
  position = "topright",
  type = "Coefficient",
  plot_label = NULL
)
}
\arguments{
\item{Z}{Coefficient matrix with each row corresponding to the coefficients and wavelength in columns}

\item{wv}{vector of wavelengths}

\item{xlim}{vector to change the default xlim of the plots (ex xlim = c(500, 2400))}

\item{position}{Position of the legend (see base function legend for help)}

\item{type}{Name of the y axis and of the legend}

\item{plot_label}{optional plot label to include with the figure}
}
\description{
f.plot.coef
}
\author{
Julien Lamour
}
