% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.plot.spec.R
\name{f.plot.spec}
\alias{f.plot.spec}
\title{f.plot.spec}
\usage{
f.plot.spec(
  Z,
  wv,
  xlim = NULL,
  position = "topright",
  type = "Reflectance",
  plot_label = NULL,
  CI = 95
)
}
\arguments{
\item{Z}{Spectra matrix with each row corresponding to a spectra and wavelength in columns}

\item{wv}{vector of wavelengths corresponding to the column of the spectra matrix Z}

\item{xlim}{vector to change the default xlim of the plots (ex xlim = c(500, 2400))}

\item{position}{Position of the legend (see base function legend for help)}

\item{type}{Name of the y axis and of the legend. E.g. Reflectance, Transmittance}

\item{plot_label}{optional plot label to include with the figure}

\item{CI}{Desired confidence interval for the spectra plot. Options are: 95
or 90. Default is: 95}
}
\description{
f.plot.spec
}
\author{
Julien Lamour, Shawn P. Serbin, Andrés Baresch
}
