% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_optimal_components.R
\name{find_optimal_comp_by_groups}
\alias{find_optimal_comp_by_groups}
\title{Uses the firstMin and firstPlateau methods for the determination of the optimal number of PLSR model components,
by group (i.e. optimal selection by stratification)}
\usage{
find_optimal_comp_by_groups(
  dataset = NULL,
  targetVariable = NULL,
  method = "firstPlateau",
  maxComps = 20,
  iterations = 20,
  prop = 0.7,
  random_seed = 123456789,
  group_variables = NULL
)
}
\arguments{
\item{dataset}{input full PLSR dataset. Usually just the calibration dataset}

\item{targetVariable}{What object or variable to use as the Y (predictand) in the PLSR model?
Usually the "inVar" variable set at the beginning of a PLS script}

\item{method}{Which approach to use to find optimal components. Options: firstPlateau, firstMin}

\item{maxComps}{maximum number of components to consider}

\item{iterations}{how many different permutations to run}

\item{prop}{proportion of data to preserve for each permutation}

\item{random_seed}{random seed to use for splitting data}

\item{group_variables}{group_variables character vector of the form c("var1", "var2"..."varn")
providing the factors used for stratified sampling.}
}
\value{
nComps the optimal number of PLSR components
}
\description{
Uses the firstMin and firstPlateau methods for the determination of the optimal number of PLSR model components,
by group (i.e. optimal selection by stratification)
}
\author{
asierrl, Shawn P. Serbin
}
