% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ecosis_data.R
\name{get_ecosis_data}
\alias{get_ecosis_data}
\title{Function to pull data from EcoSIS using the EcoSIS API}
\usage{
get_ecosis_data(ecosis_id = NULL)
}
\arguments{
\item{ecosis_id}{the alphanumeric EcoSIS API dataset ID}
}
\value{
EcoSIS spectral dataset object
}
\description{
Function to pull data from EcoSIS using the EcoSIS API
}
\examples{
\dontrun{ 
ecosis_id <- "960dbb0c-144e-4563-8117-9e23d14f4aa9"
dat_raw <- get_ecosis_data(ecosis_id = ecosis_id)
head(dat_raw)
names(dat_raw)[1:40]
}

}
\author{
Shawn P. Serbin, Alexey Shiklomanov
}
