% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{percent_rmse}
\alias{percent_rmse}
\title{Calculate RMSE and percent RMSE with PLSR model results}
\usage{
percent_rmse(
  plsr_dataset = NULL,
  inVar = NULL,
  residuals = NULL,
  range = "full"
)
}
\arguments{
\item{plsr_dataset}{input plsr dataset}

\item{inVar}{the trait variable used in the calculation of RMSE}

\item{residuals}{predicted minus observed residual vector from either
a cross-validation CV or independent validation}

\item{range}{calculate over the full data range or the 95\% of data range.
options full or 95perc}
}
\value{
output a list containing the rmse and perc_rmse.
output <- list(rmse = rmse, perc_rmse = perc_rmse)
}
\description{
Calculate RMSE and percent RMSE with PLSR model results
}
\author{
Shawn P. Serbin
}
