% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_permutation.R
\name{pls_permutation}
\alias{pls_permutation}
\title{Run a PLSR model permutation analysis. Can be used to determine the optimal number of components
or conduct a boostrap uncertainty analysis}
\usage{
pls_permutation(
  dataset = NULL,
  maxComps = 20,
  iterations = 20,
  prop = 0.7,
  verbose = FALSE
)
}
\arguments{
\item{dataset}{input full PLSR dataset. Usually just the calibration dataset}

\item{maxComps}{maximum number of components to use for each PLSR fit}

\item{iterations}{how many different permutations to run}

\item{prop}{proportion of data to preserve for each permutation}

\item{verbose}{Should the function report the current iteration status/progress to the terminal
or run silently? TRUE/FALSE. Default FALSE}
}
\description{
See Serbin et al. (2019). DOI: https://doi.org/10.1111/nph.16123
}
\author{
Julien Lamour, Shawn P. Serbin
}
