% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{source_GitHubData}
\alias{source_GitHubData}
\title{Function to source text data from GitHub}
\usage{
source_GitHubData(url, sep = ",", header = TRUE)
}
\arguments{
\item{url}{http/https URL to the github dataset}

\item{sep}{dataset file delimiter}

\item{header}{TRUE/FALSE does the file have a column header?}
}
\description{
Function to source text data from GitHub
}
\author{
gist.github.com/christophergandrud/4466237
}
