% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{testForPackage}
\alias{testForPackage}
\title{Function to check for installed package
not presently used}
\usage{
testForPackage(pkg)
}
\description{
Function to check for installed package
not presently used
}
