% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vip.R
\name{VIP}
\alias{VIP}
\title{VIP returns all VIP values for all variables and all number of components, as a ncomp x nvars matrix.}
\usage{
VIP(object)
}
\arguments{
\item{object}{fitted pls::plsr object}
}
\description{
VIP returns all VIP values for all variables and all number of components, as a ncomp x nvars matrix.
}
