% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vip.R
\name{VIPjh}
\alias{VIPjh}
\title{VIPjh returns the VIP of variable j with h components}
\usage{
VIPjh(object, j, h)
}
\arguments{
\item{object}{fitted pls::plsr object}

\item{j}{which variable in the fitted pls::plsr object}

\item{h}{the number of components in the fitted pls::plsr object to calculate the VIP}
}
\description{
VIPjh returns the VIP of variable j with h components
}
