% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{testForPackage}
\alias{testForPackage}
\title{Function to check for installed package}
\usage{
testForPackage(pkg)
}
\arguments{
\item{pkg}{name of package to check if installed
not presently used}
}
\description{
Function to check for installed package
}
